+++ templates/dart_template.cpp
@@ -57,6 +57,8 @@
 #include <limbo/kernel/squared_exp_ard.hpp>
 #include <limbo/mean/constant.hpp>
 #include <limbo/model/gp.hpp>
+#include <limbo/model/multi_gp.hpp>
+#include <limbo/model/multi_gp/parallel_lf_opt.hpp>
 #include <limbo/opt/cmaes.hpp>
 
 #include <robot_dart/position_control.hpp>
@@ -69,8 +71,6 @@
 #include <blackdrops/blackdrops.hpp>
 #include <blackdrops/gp_model.hpp>
 #include <blackdrops/model/gp/kernel_lf_opt.hpp>
-#include <blackdrops/model/multi_gp.hpp>
-#include <blackdrops/model/multi_gp/multi_gp_parallel_opt.hpp>
 #include <blackdrops/system/dart_system.hpp>
 
 // TO-CHANGE (optional): You can include other policies as well (GP and linear policy already implemented)
@@ -79,9 +79,9 @@
 // TO-CHANGE (optional): You can include other reward types as well (GPReward is already implemented)
 #include <blackdrops/reward/reward.hpp>
 
-#include <utils/cmd_args.hpp>
-#include <utils/dart_utils.hpp>
-#include <utils/utils.hpp>
+#include <blackdrops/utils/cmd_args.hpp>
+#include <blackdrops/utils/dart_utils.hpp>
+#include <blackdrops/utils/utils.hpp>
 
 struct Params {
     struct blackdrops : public ::blackdrops::defaults::blackdrops {
@@ -323,7 +323,7 @@ int main(int argc, char** argv)
     using kernel_t = limbo::kernel::SquaredExpARD<Params>;
     using mean_t = limbo::mean::Constant<Params>;
 
-    using GP_t = blackdrops::model::MultiGP<Params, limbo::model::GP, kernel_t, mean_t, blackdrops::model::multi_gp::MultiGPParallelLFOpt<Params, blackdrops::model::gp::KernelLFOpt<Params>>>;
+    using GP_t = limbo::model::MultiGP<Params, limbo::model::GP, kernel_t, mean_t, limbo::model::multi_gp::ParallelLFOpt<Params, blackdrops::model::gp::KernelLFOpt<Params>>>;
 
     using MGP_t = blackdrops::GPModel<Params, GP_t>;
